requirements.txt
Flask==3.0.3
streamlit==1.35.0
openai==1.35.13
yagmail==0.15.293
mysql-connector-python==8.4.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.1